feat: sanitize markup
