+++ packages/x6/src/util/string/string.ts
@@ -18,4 +18,5 @@ export * from './case'
 export * from './hashcode'
 export * from './uuid'
 export * from './text'
+export * from './html'
 export { exec as eval } from './eval'
