+++ packages/x6/src/util/string/text.ts
@@ -1,4 +1,3 @@
-import $ from 'jquery'
 import { v } from '../../v'
 import { Size } from '../../types'
 import { NumberExt } from '../number'
@@ -256,48 +255,4 @@ export function breakText(
  */
 export function sanitizeText(text: string) {
   return text.replace(/ /g, '\u00A0')
-}
-
-/**
- * Sanitizes HTML with jQuery prevent Application from XSS attacks.
- * ref: https://gist.github.com/ufologist/5a0da51b2b9ef1b861c30254172ac3c9
- */
-export function sanitizeHTML(html: string) {
-  // If documentContext (second parameter) is not specified or given as
-  // `null` or `undefined`, a new document is used. Inline events will not
-  // execute when the HTML is parsed; this includes, for example, sending
-  // GET requests for images.
-
-  // If keepScripts (last parameter) is `false`, scripts are not executed.
-  const output = $($.parseHTML(`<div>${html}</div>`, null, false))
-
-  output
-    .find('*')
-    .toArray()
-    .forEach(currentNode => {
-      const node = (currentNode as any) as Element
-      if (node) {
-        const attrs = node.attributes
-        if (attrs) {
-          for (let i = 0, l = attrs.length; i < l; i += 1) {
-            const attr = attrs.item(i)
-            if (attr) {
-              const val = attr.value
-              const name = attr.name
-
-              // Remove attribute names that start with "on" (e.g. onload, onerror...).
-              // Remove attribute values that start with "javascript:" pseudo protocol (e.g. `href="javascript:alert(1)"`).
-              if (
-                name.indexOf('on') === 0 ||
-                val.indexOf('javascript:') === 0
-              ) {
-                node.removeAttribute(name)
-              }
-            }
-          }
-        }
-      }
-    })
-
-  return output.html()
 }
